feat: Migrate from Java 8 to Java 17 using GitHub actions
Migrate from OpenJDK 8 using javah to javac -h. The JNI plugin is removed and the maven compile plugin is used instead.
I've added an ant task to move the generated header file to the right place.

The build bash scripts are replaced with a
Makefile which has building and packaging goals. The version number for packages is driven by the version number of the
pom.xml.

Debian and RPM packages have a custom packaging version suffix which can be set with make BUILD_NUMBER=1.
Publish RPM and DEB packages to the public Cloudsmith repository https://cloudsmith.io/~bluebird/repos/common/packages/.
The GitHub action build number is used as the iteration for the RPM and DEB packages to make them unique.

Resolve: #1